ai-services/chunking_agent/requirements.txt
 pika
 boto3
 anthropic[bedrock]
 python-dotenv
 docling
 pydantic
 langchain